simplify build step
Soon we will deploy the `app` image to docker hub. But the current setup
will fail since we have a top-level await that gets ran at the build
step (`await client.connect()`). This is problematic since we might not
have a database instance during build step (because it happens through
github actions).

This commit fixes it, by putting `client.connect` on
`src/hooks.server.ts`, and handling potential other issues. It also
removes the need for passing the environment variables on the build
image.